+++ src/components/layout/index.ts
@@ -2,4 +2,3 @@ export { default as Container } from './Container'
 export { default as Footer } from './Footer'
 export { default as Masthead } from './Masthead'
 export { default as MastheadNav } from './MastheadNav'
-export { default as PageWrapper } from './PageWrapper'
