+++ src/templates/project.tsx
@@ -6,7 +6,7 @@ import { graphql } from 'gatsby'
 import { SiteMetadata, HCardIcon } from '../types/gatsby'
 import { ProjectNode } from '../types/nodes'
 
-import { Container, PageWrapper } from '../components/layout'
+import { Container } from '../components/layout'
 import {
   PageHeader,
   PageTitle,
@@ -18,6 +18,7 @@ import {
 } from '../components/page'
 import { ProjectCard, ProjectFooter } from '../components/projects'
 import { AnchorButton } from '../components/chungking-core'
+import { PageWrapper } from '../layouts'
 
 interface ProjectTemplateProps extends RouterProps {
   data: {
